+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 - Added support for knapsack constraints
 - Added isPositive(), isNegative(), isFeasLE(), isFeasLT(), isFeasGE(), isFeasGT(), isHugeValue(), and tests
-- Added SCIP_LOCKTYPE, addVarLocksType
+- Added SCIP_LOCKTYPE, addVarLocksType, getNLocksDown, getNLocksUp, getNLocksUpType, getNLocksDownType, and tests
 ### Fixed
 ### Changed
 ### Removed